requirements.txt
fastapi
python-dotenv
uvicorn
Jinja2
langchain-google-genai
langchain-core
langdetect
transformers
torch
soundfile
librosa
langid
tf-keras
